requirements.txt
langchain>=0.1.0
openai>=1.3.0
pyautogen>=0.2.3
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
langchain-community
langchain-openai